refactor(ssl): To different file, doc and to robot.py
