+++ robot_skills/src/robot_skills/robot.py
@@ -7,6 +7,7 @@ import base
 import torso
 import arms
 import head
+import ssl
 
 # Human Robot Interaction
 import speech
@@ -42,6 +43,7 @@ class Robot(object):
         self.parts['leftArm'] = arms.Arm(self.robot_name, self.tf_listener, side="left")
         self.parts['rightArm'] = arms.Arm(self.robot_name, self.tf_listener, side="right")
         self.parts['head'] = head.Head(self.robot_name, self.tf_listener)
+        self.parts['ssl'] = ssl.SSL('/{}/ssl/direction_of_arrival'.format(self.robot_name))
 
         # Human Robot Interaction
         self.parts['lights'] = lights.Lights(self.robot_name, self.tf_listener)
